+++ packages/serverless-components/nextjs-component/src/component.ts
@@ -198,7 +198,10 @@ class NextjsComponent extends Component {
           logLambdaExecutionTimes: inputs.logLambdaExecutionTimes || false,
           domainRedirects: inputs.domainRedirects || {},
           minifyHandlers: inputs.minifyHandlers || false,
-          enableHTTPCompression: false
+          enableHTTPCompression: false,
+          handler: inputs.handler
+            ? `${inputs.handler.split(".")[0]}.js`
+            : undefined
         },
         nextStaticPath
       );
@@ -235,6 +238,7 @@ class NextjsComponent extends Component {
       priceClass: cloudFrontPriceClassInputs,
       errorPages: cloudFrontErrorPagesInputs,
       distributionId: cloudFrontDistributionId = null,
+      comment: cloudFrontComment,
       ...cloudFrontOtherInputs
     } = inputs.cloudfront || {};
 
@@ -409,7 +413,7 @@ class NextjsComponent extends Component {
         description: inputs.description
           ? `${inputs.description} (API)`
           : "API Lambda@Edge for Next CloudFront distribution",
-        handler: "index.handler",
+        handler: inputs.handler || "index.handler",
         code: join(nextConfigPath, API_LAMBDA_CODE_DIR),
         role: {
           service: ["lambda.amazonaws.com", "edgelambda.amazonaws.com"],
@@ -457,7 +461,7 @@ class NextjsComponent extends Component {
       description:
         inputs.description ||
         "Default Lambda@Edge for Next CloudFront distribution",
-      handler: "index.handler",
+      handler: inputs.handler || "index.handler",
       code: join(nextConfigPath, DEFAULT_LAMBDA_CODE_DIR),
       role: {
         service: ["lambda.amazonaws.com", "edgelambda.amazonaws.com"],
@@ -582,7 +586,8 @@ class NextjsComponent extends Component {
       }),
       ...(cloudFrontErrorPagesInputs && {
         errorPages: cloudFrontErrorPagesInputs
-      })
+      }),
+      comment: cloudFrontComment
     });
 
     let appUrl = cloudFrontOutputs.url;