+++ packages/libs/lambda-at-edge/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.7.0-alpha.17](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/lambda-at-edge@1.7.0-alpha.16...@sls-next/lambda-at-edge@1.7.0-alpha.17) (2020-10-21)
+
+### Bug Fixes
+
+- **lambda-at-edge:** non-dynamic routes for rewrite should not include routes in prerender manifest ([#700](https://github.com/serverless-nextjs/serverless-next.js/issues/700)) ([06d4edb](https://github.com/serverless-nextjs/serverless-next.js/commit/06d4edb72da2662affc9ad25a2e4eb5550d415e9))
+- **lambda-at-edge:** serve HTML pages with no props (i.e static pages) properly on preview mode enabled ([#701](https://github.com/serverless-nextjs/serverless-next.js/issues/701)) ([57cd668](https://github.com/serverless-nextjs/serverless-next.js/commit/57cd6689152530064635eb9d53c5201d580c9b61))
+
+# [1.7.0-alpha.16](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/lambda-at-edge@1.7.0-alpha.15...@sls-next/lambda-at-edge@1.7.0-alpha.16) (2020-10-21)
+
+### Bug Fixes
+
+- **lambda-at-edge, e2e-tests:** don't rewrite non-dynamic api routes ([#698](https://github.com/serverless-nextjs/serverless-next.js/issues/698)) ([27526e7](https://github.com/serverless-nextjs/serverless-next.js/commit/27526e70c57d3bab016f734a0672d62c9d43bcca))
+
+# [1.7.0-alpha.15](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/lambda-at-edge@1.7.0-alpha.14...@sls-next/lambda-at-edge@1.7.0-alpha.15) (2020-10-20)
+
+### Bug Fixes
+
+- **lambda-at-edge, e2e-tests:** non-dynamic routes should never be rewritten ([#697](https://github.com/serverless-nextjs/serverless-next.js/issues/697)) ([f16c9a0](https://github.com/serverless-nextjs/serverless-next.js/commit/f16c9a0811e80bb5d45d9ebd517c2ae8f3b35316))
+
+# [1.7.0-alpha.14](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/lambda-at-edge@1.7.0-alpha.13...@sls-next/lambda-at-edge@1.7.0-alpha.14) (2020-10-20)
+
+### Features
+
+- **lambda-at-edge, nextjs-component:** allow `handler` input for custom handler code ([#649](https://github.com/serverless-nextjs/serverless-next.js/issues/649)) ([cecd327](https://github.com/serverless-nextjs/serverless-next.js/commit/cecd327993d593c582eeb140538b271b2eef509a))
+
 # [1.7.0-alpha.13](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/lambda-at-edge@1.7.0-alpha.12...@sls-next/lambda-at-edge@1.7.0-alpha.13) (2020-10-18)
 
 **Note:** Version bump only for package @sls-next/lambda-at-edge