+++ packages/serverless-components/domain/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-alpha.3](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/domain@1.1.0-alpha.2...@sls-next/domain@1.1.0-alpha.3) (2020-10-20)
+
+### Bug Fixes
+
+- **serverless-component:** fix for AWS accounts with more than 100 domains ([69574c7](https://github.com/serverless-nextjs/serverless-next.js/commit/69574c7d82bee01817bceb0f9bb7e8b9cae53faf))
+
 # [1.1.0-alpha.2](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/domain@1.1.0-alpha.1...@sls-next/domain@1.1.0-alpha.2) (2020-10-07)
 
 **Note:** Version bump only for package @sls-next/domain